some feed timing updates
