+++ feedreader/lib/feedreader_web/router.ex
@@ -56,31 +56,22 @@ defmodule FeedreaderWeb.Router do
       reset_path: "/reset",
       auth_routes_prefix: "/auth",
       on_mount: [{FeedreaderWeb.LiveUserAuth, :live_no_user}],
-      overrides: [
-        FeedreaderWeb.AuthOverrides,
-        Elixir.AshAuthentication.Phoenix.Overrides.DaisyUI
-      ]
+      overrides: [AshAuthentication.Phoenix.Overrides.DaisyUI]
     )
 
-    # Remove this if you do not want to use the reset password feature
     reset_route(
       auth_routes_prefix: "/auth",
-      overrides: [
-        FeedreaderWeb.AuthOverrides,
-        Elixir.AshAuthentication.Phoenix.Overrides.DaisyUI
-      ]
+      overrides: [AshAuthentication.Phoenix.Overrides.DaisyUI]
     )
 
-    # Remove this if you do not use the confirmation strategy
     confirm_route(Feedreader.Accounts.User, :confirm_new_user,
       auth_routes_prefix: "/auth",
-      overrides: [FeedreaderWeb.AuthOverrides, Elixir.AshAuthentication.Phoenix.Overrides.DaisyUI]
+      overrides: [AshAuthentication.Phoenix.Overrides.DaisyUI]
     )
 
-    # Remove this if you do not use the magic link strategy.
     magic_sign_in_route(Feedreader.Accounts.User, :magic_link,
       auth_routes_prefix: "/auth",
-      overrides: [FeedreaderWeb.AuthOverrides, Elixir.AshAuthentication.Phoenix.Overrides.DaisyUI]
+      overrides: [AshAuthentication.Phoenix.Overrides.DaisyUI]
     )
   end
