+++ feedreader/lib/feedreader_web/telemetry.ex
@@ -9,11 +9,7 @@ defmodule FeedreaderWeb.Telemetry do
   @impl true
   def init(_arg) do
     children = [
-      # Telemetry poller will execute the given period measurements
-      # every 10_000ms. Learn more here: https://hexdocs.pm/telemetry_metrics
       {:telemetry_poller, measurements: periodic_measurements(), period: 10_000}
-      # Add reporters as children of your supervision tree.
-      # {Telemetry.Metrics.ConsoleReporter, metrics: metrics()}
     ]
 
     Supervisor.init(children, strategy: :one_for_one)
@@ -21,7 +17,6 @@ defmodule FeedreaderWeb.Telemetry do
 
   def metrics do
     [
-      # Phoenix Metrics
       summary("phoenix.endpoint.start.system_time",
         unit: {:native, :millisecond}
       ),
@@ -51,8 +46,6 @@ defmodule FeedreaderWeb.Telemetry do
         tags: [:event],
         unit: {:native, :millisecond}
       ),
-
-      # Database Metrics
       summary("feedreader.repo.query.total_time",
         unit: {:native, :millisecond},
         description: "The sum of the other measurements"
@@ -74,8 +67,6 @@ defmodule FeedreaderWeb.Telemetry do
         description:
           "The time the connection spent waiting before being checked out for the query"
       ),
-
-      # VM Metrics
       summary("vm.memory.total", unit: {:byte, :kilobyte}),
       summary("vm.total_run_queue_lengths.total"),
       summary("vm.total_run_queue_lengths.cpu"),
@@ -84,10 +75,6 @@ defmodule FeedreaderWeb.Telemetry do
   end
 
   defp periodic_measurements do
-    [
-      # A module, function and arguments to be invoked periodically.
-      # This function must call :telemetry.execute/3 and a metric must be added above.
-      # {FeedreaderWeb, :count_users, []}
-    ]
+    []
   end
 end
